multi-node ESX support in tenancy/auth code [post 0.13] (#1065)
Implemented multi-ESX support for auth.db. See ConfigDB_State.md and admin-cl.md files in this commit

- refactored AuthenticationDataManager to deal with modes and rearranged there a lot
- added all commands to CLI and related docs.
- added doc and (some) tests

Also to make tests pass
* Replaced remaining vm-group with vmgroup term

+++ esx_service/utils/auth.py
@@ -50,10 +50,10 @@ def get_auth_mgr():
 
 def get_default_tenant():
     """
-        Get DEFAULT tenant by querying the auth DB.
+        Get DEFAULT tenant by querying the auth DB or from hardcoded defaults.
         VM which does not belong to any tenant explicitly will
         be assigned to DEFAULT tenant if DEFAULT tenant exists
-        Return value:
+        Returns: (err, uuid, name)
         -- error_msg: return None on success or error info on failure
         -- tenant_uuid: return DEFAULT tenant uuid on success,
            return None on failure or DEFAULT tenant does not exist
@@ -64,6 +64,10 @@ def get_default_tenant():
     err_msg, _auth_mgr = get_auth_mgr()
     if err_msg:
         return err_msg, None, None
+
+    if _auth_mgr.allow_all_access():
+        return None, auth_data_const.DEFAULT_TENANT_UUID, auth_data_const.DEFAULT_TENANT
+
     try:
         cur = _auth_mgr.conn.execute(
             "SELECT id FROM tenants WHERE name = ?",
@@ -100,6 +104,10 @@ def get_default_privileges():
     err_msg, _auth_mgr = get_auth_mgr()
     if err_msg:
         return err_msg, None
+
+    if _auth_mgr.allow_all_access():
+        return None, _auth_mgr.get_default_privileges_dict()
+
     privileges = []
     logging.debug("get_default_privileges")
 
@@ -119,7 +127,7 @@ def get_default_privileges():
 def get_tenant(vm_uuid):
     """
         Get tenant which owns this VM by querying the auth DB.
-        Return value:
+        Return: error_msg, tenant_uuid, tenant_name
         -- error_msg: return None on success or error info on failure
         -- tenant_uuid: return tenant uuid which the VM with given vm_uuid is associated to,
            return None if the VM is not associated to any tenant
@@ -129,10 +137,15 @@ def get_tenant(vm_uuid):
     err_msg, _auth_mgr = get_auth_mgr()
     if err_msg:
         return err_msg, None, None
+    logging.debug("auth.get_tenant: allow_all: %s, uuid: %s", _auth_mgr.allow_all_access(), vm_uuid)
+    if _auth_mgr.allow_all_access():
+        logging.debug("returning default info")
+        return None, auth_data_const.DEFAULT_TENANT_UUID, auth_data_const.DEFAULT_TENANT
+
     try:
         cur = _auth_mgr.conn.execute(
-                    "SELECT tenant_id FROM vms WHERE vm_id = ?",
-                    (vm_uuid, )
+            "SELECT tenant_id FROM vms WHERE vm_id = ?",
+            (vm_uuid, )
         )
         result = cur.fetchone()
     except sqlite3.Error as e:
@@ -185,6 +198,9 @@ def get_privileges(tenant_uuid, datastore_url):
     err_msg, _auth_mgr = get_auth_mgr()
     if err_msg:
         return err_msg, None
+
+    # TBD return _auth_mgr.get_defaut_privileges_dict() when both params matchm otherwise return std err
+
     privileges = []
     logging.debug("get_privileges tenant_uuid=%s datastore_url=%s", tenant_uuid, datastore_url)
     try:
@@ -406,17 +422,18 @@ def authorize(vm_uuid, datastore_url, cmd, opts):
     logging.debug("Authorize: cmd=%s", cmd)
     logging.debug("Authorize: opt=%s", opts)
 
-    try:
-        get_auth_mgr()
-    except (auth_data.DbConnectionError, auth_data.DbAccessError) as e:
-        error_msg = str(e)
+    error_msg, _auth_mgr = get_auth_mgr()
+    if error_msg:
         return error_msg, None, None
+
+    if _auth_mgr.allow_all_access():
+        return None, auth_data_const.DEFAULT_TENANT_UUID, auth_data_const.DEFAULT_TENANT
 
     # If table "tenants", "vms", "privileges" or "volumes" does not exist
     # don't need auth check
     if not tables_exist():
-        logging.error("Required tables in auth db do not exist")
-        error_msg = "Required tables in aut db do not exist"
+        error_msg = "Required tables do not exist in auth db"
+        logging.error(error_msg)
         return error_msg, None, None
 
     error_msg, tenant_uuid, tenant_name = get_tenant(vm_uuid)
@@ -457,6 +474,14 @@ def add_volume_to_volumes_table(tenant_uuid, datastore_url, vol_name, vol_size_i
     logging.debug("add to volumes table(%s %s %s %s)", tenant_uuid, datastore_url,
                   vol_name, vol_size_in_MB)
 
+    if _auth_mgr.allow_all_access():
+        logging.debug("Skipping Add volume to DB %s (allow_all_access)", tenant_uuid)
+        if tenant_uuid != auth_data_const.DEFAULT_TENANT_UUID:
+            return _auth_mgr.err_config_init_needed()
+        else:
+            logging.info("No access control, skipping volumes tracing in auth DB")
+            return None
+
     try:
         _auth_mgr.conn.execute(
             "INSERT INTO volumes(tenant_id, datastore_url, volume_name, volume_size) VALUES (?, ?, ?, ?)",
@@ -481,6 +506,11 @@ def remove_volume_from_volumes_table(tenant_uuid, datastore_url, vol_name):
 
     logging.debug("remove volumes from volumes table(%s %s %s)", tenant_uuid, datastore_url,
                   vol_name)
+
+    if _auth_mgr.allow_all_access():
+        logging.debug("Skipping Rm volume from DB %s (allow_all_access)", tenant_uuid)
+        return None
+
     try:
         _auth_mgr.conn.execute(
                     "DELETE FROM volumes WHERE tenant_id = ? AND datastore_url = ? AND volume_name = ?",
